# DRAFT: Lazy well trajectory