--- a/src/xtgeo/xyz/_xyz_data.py
+++ b/src/xtgeo/xyz/_xyz_data.py
@@ -324,6 +324,8 @@ class _XYZData:
 
         # the purpose of this hash check is to avoid spending time on consistency
         # checks if no changes
+        
+        return
         hash_proposed = (
             jhash(self._df),
             jhash(self._attr_types),